design: A2hsSnackbar 스타일 적용
